fixes issue 35
Added class DirtyHack, because I want to keep my dirty hacks in a central place.
Moved IMessagesResource to the translator.rbe plugin, to make it accessible for 
babel.core. Moved also the Interface it needed. Anyway, the central thing I did to
fix this issue was: New API in RBManager: writeFile with whole BundleGroup as param
or just a Bundle. This should be invoked from the RB-View if there is write-access. This
API writes directly to the IFile, therefore the Editor won't be dirty! Well, not just because of
that, I added another hack to disable the setText method, which is called when something
gets serialized. Have FUN!