performance improvement for $NON-NLS$ line comment search
Instead of using AST and searching through all comments of document, now only using Document object to get whole line of a StringLiteral and using String operations to search for a line comment at at end of line. 
The problem with AST and CompilationUnit.getLastTrailingCommentIndex() method is, that it is only applicable with AST structure. So I could only get a comment, for example, at the end of a statement instead of the end of a line.