fixes issue 37
IMessagesEditorListener got a new method, which tells the listener which resource has been changed.
Now it is possible for the I18NPage to know which key it has to update. Prior to this change, the update was triggered by a change of the selecion of the key in the key-tree-view of the editor.