fixes issue 36
This was a very complex change. The goal was, that the RB-View can function without an open editor in the background. For this to accomplish I examined the behavior of the ResourceBundlemanager and tried detect and instantiate resource bundles on demand. The tricky thing was to swap the property messages bundles with the eclipse messages bundles, so that the editor can function properly. To achieve that, I made some improvements in notifyMessagesBundleCreated in the RBManager. I also modified the dispose() function of the PropertiesFileResource, because the FileMonitor caused an out of sync, because it triggered a setText after the save. Another tricky thing was to detect the "Async-ness" of the editor and the RB-View, when the swap was finished. I noticed, that ResKeyTreeContentProvider worked with an old copy of the messages bundle group -> anyway. Fixed that too. Another fix was, to ensure, that the build.properties would not appear in the Selection of the editable resource bundles. So long