ResourceBundleWizard now checks if the list of Locales is empty + usability improvement
Fixes Issue 47

The "Finish" button will now be automaticly greyed out and display a warning message at the top if the user has not yet added any Locales to the list.
To not obscure the instructions at the top with a warning and improve usability there will now be at least the "[Default]" locale in the list when the ResourceBundleWizard is launched.