Fixed ResourceBundleWizard file/folder-creation
Fixes Issue 54
"Finish" button will be greyed out when:
- an invalid folder-path has been entered
- the folder-path is in a non-existing project
If you click "Finish" and the folder does not exist, you will be asked if you want to create it. Cancelling this will take  you back to the wizard. Click OK and the folder (or folderstructure) will be created before the file is created.
You should not get any meaningless error-messages anymore