The Treeview in the ResourceBundle editor now has tooltips
Fixes Issue 49
If a key in the TreeView of the ResourceBundle has missing translations or has duplicate values a tooltip will now be shown which tells the user about this warning/error. Till now the user had to open the key and check manually what kind of error/warning the key has.