Icons in the Message Editor now use decorators + transparent images
Fixes Issue 65
UIUtils now provide methods to get Images of the key with either a warning or error icon depending on what issue the key has. IN THEORY the issue of duplicate values being used in multiple keys should work aswell. The key will be greyed out then. Though this feature is currently non-existent or seriously bugged
I also edited some of the icons to be transparent.