[RAP] changed code to Single Source
Most of the code gets now shared for RAP + RCP Translator and TapiJI Tools.
Created own fragments for RAP and RCP, which store the different implementation for running the Translator as a RAP web application and a RCP standalone Eclipse application and for the TapiJI Tools Eclipse Extension (needs the rcp fragments to run).
Also created compatility plugins, which extend the RAP tools in rap.compat and just delegate in the rcp.compat packages.
Renamed extResources to supplemental, containing classes (mostly only parts) needed for babel and tapiji to run, but not exist in RAP target platform.