[RAP] added downloading action in file menu
Added a downloading action as menu contribution, to be able to download properties files to the client.
The action creates a download url with the filename and the download service handler as parameters and opens this url with the browser widget. The download service handler takes incoming download requests and streams the content of requested file to the client.