[RAP] added login, logout functionality and a Storage View
Implemented dialogs, menu entries and actions for user login and logout, and also the logic behind (verify username and password, ...).
The user is stored in the RWT session as long as he/she is logged in and the session is valid.
Using a context variable to hide/show menu entries (login/logout), which gets set properly after user performs a login/logout.
The Storage View shows the stored files of a logged in user. It has 3 different "views": If the user isn't logged in, it shows links to login or to register (TODO). If the user has no files in storage, it shows a link to upload a file. And if the user is logged in and has stored files it shows them (TODO).
Removed Glossary View, because its not working with current file managment -> needs to be adapted (upload, download).