implemented "add new local" action
fixes issue 55

Created simple dialog, which only holds a LocaleSelector composite to choose the new locale. Existing locals will get ignored. Calling the MessagesGroupBundle.addMessagesBundle(Locale) method which delegates to the IMessagesBundleGroupStrategy.createMessagesBundle(Locale) method.
Implemented the createMessagesBundle(Locale) method in the DefaultBundleGroupStrategy, which simple creates a new IFile in the location of the already opened file and delegates the created resource to the createMessagesBundle(Locale,IResource) method.
Extracted a addMessagesBundle(MessagesBundle, Locale) method in MessagesEditor and a addI18NEntry(MessagesEditor, Locale) in I18NPage. Added an "addMesssagesBundle" Listener to the messages bundle group which is hold by the MessagesEditor. The Listener calls the extracted methods which add a new bundle to the editor (new text editor for locale) and creates a new i18n entry in the i18n page.