[RAP] fixes issue 93 + minor clean up
Fix: The user object was saved in the RAP session, which is deleted when the page is refreshed. Now the user object is saved in the underlying HTTP session, which survives the refresh. Additionally the context variable needs to be reseted after the refresh.
Clean up: Removed the duplicated UserUtils class.