[RAP] disposing messages bundle group when editor is closed
fixes issue 100

When the RBManager is initialized (= when MessagesEditor is opened the first time), it creates (passive loading) a new messages bundle group (MBG) with a PropertiesFileResource, which actives a TimerTask that recreates the underlying files. So this MBG has to be dispose to remove the TimerTask. This is done, when the editor is closed.